pyproject.toml
[project]
name = "drawtool"
version = "0.1.0"
description = "PowerPoint-like figure builder referencing external assets by path"
requires-python = ">=3.11"
dependencies = ["Pillow>=10.0.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
